Connect to a Presto database

Connect to a Presto database

Usage

Presto(...)

# S4 method for PrestoDriver
dbConnect(
  drv,
  catalog,
  schema,
  user,
  host = "localhost",
  port = 8080,
  source = methods::getPackageName(),
  session.timezone = "UTC",
  parameters = list(),
  request.config = httr::config(),
  use.trino.headers = FALSE,
  extra.credentials = "",
  bigint = c("integer", "integer64", "numeric", "character"),
  ...
)

# S4 method for PrestoConnection
dbDisconnect(conn)

Arguments

...

currently ignored

drv

A driver object generated by Presto

catalog

The catalog to be used

schema

The schema to be used

user

The current user

host

The presto host to connect to

port

Port to use for the connection

source

Source to specify for the connection

session.timezone

Time zone to use for the connection. Presto returns timestamps without time zones with respect to this value. The time arithmetic (e.g. adding hours) will also be done in the given time zone. See the session.timezone tests for examples.

parameters

A list of extra parameters to be passed in the ‘X-Presto-Session’ header

request.config

An optional config list, as returned by httr::config(), to be sent with every HTTP request.

use.trino.headers

A boolean to indicate whether Trino request headers should be used. Default to FALSE.

extra.credentials

Extra credentials to be passed in the X-Presto-Extra-Credential or X-Trino-Extra-Credential header ( depending on the value of the use.trino.headers argument). Default to an empty string.

bigint

The R type that Presto's 64-bit integer (BIGINT) class should be translated to. The default is "integer", which returns R's integer type, but results in NA for values above/below +/-2147483647. "integer64" returns a bit64::integer64, which allows the full range of 64 bit integers. "numeric" coerces into R's double type but might result in precision loss. Lastly, "character" casts into R's character type.

conn

A PrestoConnection object

Value

[Presto] A PrestoDriver object

dbConnect A PrestoConnection object

[dbDisconnect] A logical value indicating success

Examples

if (FALSE) {
  conn <- dbConnect(Presto(), catalog = 'hive', schema = 'default',
                    user = 'onur', host = 'localhost', port = 8080,
                    session.timezone='US/Eastern', bigint = 'character')
  dbListTables(conn, '%_iris')
  dbDisconnect(conn)
}