S4 implementation of DBIConnection for Presto.

S4 implementation of DBIConnection for Presto.

Usage

# S4 method for PrestoConnection
show(object)

# S4 method for PrestoConnection
dbCreateTable(
  conn,
  name,
  fields,
  with = NULL,
  ...,
  row.names = NULL,
  temporary = FALSE
)

# S4 method for PrestoConnection
dbCreateTableAs(conn, name, statement, with = NULL, ...)

# S4 method for PrestoConnection,character
dbExistsTable(conn, name, ...)

# S4 method for PrestoConnection,character
dbGetQuery(conn, statement, ...)

# S4 method for PrestoConnection,character
dbListFields(conn, name, ...)

# S4 method for PrestoConnection
dbListTables(conn, pattern, ...)

# S4 method for PrestoConnection,character
dbRemoveTable(conn, name, ..., fail_if_missing = TRUE)

# S4 method for PrestoConnection,character
dbSendQuery(conn, statement, ...)

# S4 method for PrestoConnection
sqlCreateTable(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  with = NULL,
  ...
)

# S4 method for PrestoConnection
sqlCreateTableAs(con, table, statement, with = NULL, ...)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

fields

Either a character vector or a data frame.

A named character vector: Names are column names, values are types. Names are escaped with dbQuoteIdentifier(). Field types are unescaped.

A data frame: field types are generated using dbDataType().

with

An optional WITH clause for the CREATE TABLE statement.

...

Other parameters passed on to methods.

row.names

Must be NULL.

temporary

If TRUE, will generate a temporary table statement.

statement

a character string containing SQL.

pattern

optional SQL pattern for filtering table names, e.g. ‘%test%’

fail_if_missing

If `FALSE`, `dbRemoveTable()` succeeds if the table doesn't exist.

con

A database connection.

table

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')