PrestoCursor-class {RPresto}	R Documentation

Internal implementation detail class needed for its side-effects. When dbFetch is called, we need to both return the data and update the uri to the next value.

Description

Internal implementation detail class needed for its side-effects. When dbFetch is called, we need to both return the data and update the uri to the next value.

[Package RPresto version 1.3.4 Index]