.json.tabular.to.data.frame {RPresto}	R Documentation

Convert a data.frame formatted in the list of lists style as returned by Presto to an actual data.frame

Description

It ‘defines’ a few extra “data types” to handle all types returned by Presto, like timestamps with time zones.

Usage

.json.tabular.to.data.frame(data, column.types, timezone)

Arguments

data
a list of lists

column.types
A character vector of (extended) data types corresponding to each column. See .presto.to.R for a list of possible values.

timezone
The timezone to use for date types.

Details

If the items in data are named lists the column names will be inferred from those names. Otherwise they will be paste0('X', seq_along(data[[1]]))

Value

A data.frame of length(data) rows and length(data[[1]]) columns.

Note

For NA values, data should have a NULL item in the correct spot. Ragged arrays are not supported (i.e. all sublists have to have the same length). These NULL value(s) will be replaced by NA in-place for the input data.

See Also

The corresponding unit tests for a full list of capabilities and data types supported

[Package RPresto version 1.3.4 Index]