dplyr integration to connect to a Presto database.

Allows you to connect to an existing database through a presto connection.

Use src_presto to connect to an existing database, and tbl to connect to tables within that database. If you're unsure of the arguments to pass, please ask your database administrator for the values of these variables.

src_presto(
  catalog = NULL,
  schema = NULL,
  user = NULL,
  host = NULL,
  port = NULL,
  source = NULL,
  session.timezone = NULL,
  parameters = NULL,
  bigint = c("integer", "integer64", "numeric", "character"),
  ...
)

tbl.src_presto(src, from, ...)

Arguments

catalog

Catalog to use in the connection

schema

Schema to use in the connection

user

User name to use in the connection

host

Host name to connect to the database

port

Port number to use with the host name

source

Source to specify for the connection

session.timezone

Time zone for the connection

parameters

Additional parameters to pass to the connection

bigint

The R type that Presto's 64-bit integer (`BIGINT`) types should be translated to. The default is `"integer"`, which returns R's `integer` type, but results in `NA` for values above/below +/- 2147483647. `"integer64"` returns a [bit64::integer64], which allows the full range of 64 bit integers. `"numeric"` coerces into R's `double` type but might result in precision loss. Lastly, `"character"` casts into R's `character` type.

...

For src_presto other arguments passed on to the underlying database connector dbConnect. For tbl.src_presto, it is included for compatibility with the generic, but otherwise ignored.

src

A presto src created with src_presto.

from

Either a string giving the name of table in database, or sql described a derived table or compound join.

Examples